Updates to PruneStack 3.x
This version will work with an upcoming 3.2.x or > build once the PR is made into the master branch.  Thanks to DLLLarson (Dale Larson) and others who worked on making adjustments to the chord.remove() exposed method (and the parent() read access).  This feature also adds the ability to "undo" the plugin action.  Again, Dale added in this code as I wasn't familiar with how to do that.  Finally, the "Move to Voice" functionality has been changed to "Prune & Move to Voice" - I'm not 100% satisfied with this but it's somewhat useful and I would argue more useful than the 2.x version which just moved the selected notes to the new layer.  As mentioned in the code, I wanted to move only the notes not in the selected chord levels, but that would require the node.selected method to be non-readonly.